+++ src/CreditEnhancement.hs
@@ -23,6 +23,7 @@ import Data.Fixed
 import Data.Maybe
 import Types
 import Util
+import DateUtil
 import Stmt
 
 type LiquidityProviderName = String