+++ test/UT/AssetTest.hs
@@ -322,35 +322,35 @@ armTest =
         (CF.sizeCashFlowFrame arm1_cf)
       ,testCase "ARM case 1/ first cash" $
         assertEqual "first cash row"
-        (Just (CF.MortgageFlow (L.toDate "20240501") 227.66 12.34 0.6 0 0 0 0 0 0.03 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20240501") 227.66 12.34 0.6 0 0 0 0 0.03 Nothing Nothing ))
         (CF.cfAt arm1_cf 0)
       ,testCase "ARM case 1/ frist reset" $
         assertEqual "first rate"
-        (Just (CF.MortgageFlow (L.toDate "20240601") 215.41 12.25 0.85 0 0 0 0 0 0.045 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20240601") 215.41 12.25 0.85 0 0 0 0 0.045 Nothing Nothing ))
         (CF.cfAt arm1_cf 1)
       ,testCase "ARM case 1/periodic reset " $
         assertEqual "first rate"
-        (Just (CF.MortgageFlow (L.toDate "20240801") 190.85 12.26 0.93 0 0 0 0 0 0.055 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20240801") 190.85 12.26 0.93 0 0 0 0 0.055 Nothing Nothing ))
         (CF.cfAt arm1_cf 3)
       ,testCase "ARM case 1/remains same before next reset" $
         assertEqual "period before first reset"
-        (Just (CF.MortgageFlow (L.toDate "20240901") 178.53 12.32 0.87 0 0 0 0 0 0.055 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20240901") 178.53 12.32 0.87 0 0 0 0 0.055 Nothing Nothing ))
         (CF.cfAt arm1_cf 4)
       ,testCase "ARM case 1" $
         assertEqual "reset with periodic cap"
-        (Just (CF.MortgageFlow (L.toDate "20241201") 141.47 12.38 0.96 0 0 0 0 0 0.075 Nothing Nothing))
+        (Just (CF.MortgageFlow (L.toDate "20241201") 141.47 12.38 0.96 0 0 0 0 0.075 Nothing Nothing))
         (CF.cfAt arm1_cf 7)
       ,testCase "ARM case 1" $
         assertEqual "Period 9"
-        (Just (CF.MortgageFlow (L.toDate "20250101") 129.01 12.46 0.88 0 0 0 0 0 0.075 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20250101") 129.01 12.46 0.88 0 0 0 0 0.075 Nothing Nothing ))
         (CF.cfAt arm1_cf 8)
       ,testCase "ARM case 1" $
         assertEqual "Period 10"
-        (Just (CF.MortgageFlow (L.toDate "20250201") 116.49 12.52 0.85 0 0 0 0 0 0.08 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20250201") 116.49 12.52 0.85 0 0 0 0 0.08 Nothing Nothing ))
         (CF.cfAt arm1_cf 9)
       ,testCase "ARM case 1" $
         assertEqual "life cap"
-        (Just (CF.MortgageFlow (L.toDate "20250401") 91.24 12.65 0.77 0 0 0 0 0 0.09 Nothing Nothing ))
+        (Just (CF.MortgageFlow (L.toDate "20250401") 91.24 12.65 0.77 0 0 0 0 0.09 Nothing Nothing ))
         (CF.cfAt arm1_cf 11)
 
     ]
@@ -391,58 +391,58 @@ ppyTest =
     testGroup "Prepay Penalty tests" [
       testCase "ppy case 1" $
         assertEqual " using rate0"
-        (Just (CF.MortgageFlow (L.toDate "20210201") 9589.55 384.62 66.48 25.83 0 0 0 0 0.08 Nothing (Just 2.58)))
+        (Just (CF.MortgageFlow (L.toDate "20210201") 9589.55 384.62 66.48 25.83 0 0 0 0.08 Nothing (Just 2.58)))
         (CF.cfAt ppy_cf_1 0)
       ,testCase "ppy case 1" $
         assertEqual " using rate1"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just 0.21 )))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just 0.21 )))
         (CF.cfAt ppy_cf_1 3)
       ,testCase "ppy case 2" $
         assertEqual " using fix amount"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just 100 )))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just 100 )))
         (CF.cfAt ppy_cf_2 3)
       ,testCase "ppy case 2 1_0" $
         assertEqual " using fix amount in period"
-        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0 0.08 Nothing  (Just 100 )))
+        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0.08 Nothing  (Just 100 )))
         (CF.cfAt ppy_cf_2_1 0)
       ,testCase "ppy case 2 1" $
         assertEqual " using fix amount out of period"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just 0 )))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just 0 )))
         (CF.cfAt ppy_cf_2_1 3)
       ,testCase "ppy case 3" $
         assertEqual " using life time pct"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just 0.21 )))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just 0.21 )))
         (CF.cfAt ppy_cf_3 3)
       ,testCase "ppy case 3 1_0" $
         assertEqual " using pct in period"
-        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0 0.08 Nothing  (Just 0.25 )))
+        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0.08 Nothing  (Just 0.25 )))
         (CF.cfAt ppy_cf_3_1 0)
       ,testCase "ppy case 3 1" $
         assertEqual " using pct out of period"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just 0 )))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just 0 )))
         (CF.cfAt ppy_cf_3_1 3)
       ,testCase "ppy case 4" $
         assertEqual " using slide at period 0"
-        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0 0.08 Nothing  (Just 2.58 )))
+        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0.08 Nothing  (Just 2.58 )))
         (CF.cfAt ppy_cf_4 0)
       ,testCase "ppy case 4 1" $
         assertEqual " using slide at period 1"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just (0.07*21.92))))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just (0.07*21.92))))
         (CF.cfAt ppy_cf_4 3)     
       ,testCase "ppy case 5" $
         assertEqual " using rate 0 before 2 periods"
-        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0 0.08 Nothing  (Just (25.83*0.5) )))
+        (Just (CF.MortgageFlow (L.toDate "20210201")  9589.55 384.62 66.48 25.83 0 0 0 0.08 Nothing  (Just (25.83*0.5) )))
         (CF.cfAt ppy_cf_5 0)
       ,testCase "ppy case 5 1" $
         assertEqual " using rate 1 after 2 periods"
-        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0 0.08 Nothing (Just (0.2*21.92))))
+        (Just (CF.MortgageFlow (L.toDate "20210501") 8357.98 389.45 58.31 21.92 0 0 0 0.08 Nothing (Just (0.2*21.92))))
         (CF.cfAt ppy_cf_5 3)   
     ]
 
 delinqScheduleCFTest = 
   let 
-    cfs = [CF.MortgageFlow (L.toDate "20230901") 1000  0 0 0 0 0 0 0 0.08 Nothing Nothing
-          ,CF.MortgageFlow (L.toDate "20231001") 500 500 0 0 0 0 0 0 0.08 Nothing Nothing
+    cfs = [CF.MortgageDelinqFlow (L.toDate "20230901") 1000  0 0 0 0 0 0 0 0.08 Nothing Nothing
+          ,CF.MortgageDelinqFlow (L.toDate "20231001") 500 500 0 0 0 0 0 0 0.08 Nothing Nothing
           ]
     pool = P.Pool ([]::[AB.Mortgage])
                   (Just (CF.CashFlowFrame cfs))
@@ -465,26 +465,30 @@ delinqScheduleCFTest =
       testCase "case 01" $
         assertEqual "size of cashflow" 
         7
-        (CF.sizeCashFlowFrame poolCf) 
+        (CF.sizeCashFlowFrame poolCf) -- `debug` ("\n>>>>> Pool cf from test schedule delinq\n >>>>"++ show poolCf)
+      ,testCase "case 01_Dates" $
+        assertEqual "Dates of cashflow" 
+        (L.toDate <$> ["20230901","20231001","20231031","20231130","20231231","20240131","20240229"])
+        (CF.getDatesCashFlowFrame poolCf)
       ,testCase "case 02" $
         assertEqual "first row of cf"
-        (Just (CF.MortgageFlow (L.toDate "20230901") 995.66  0 0 0 4.34 0 0 0 0.08 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20230901") 995.66  0 0 0 4.34 0 0 0 0.08 Nothing Nothing))
         (CF.cfAt poolCf 0)
       ,testCase "case 03" $
         assertEqual "second row of cf"
-        (Just (CF.MortgageFlow (L.toDate "20231001") 493.66  497.82 0 0 4.18 0 0 0 0.08 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20231001") 493.66  497.82 0 0 4.18 0 0 0 0.08 Nothing Nothing))
         (CF.cfAt poolCf 1)
       ,testCase "case 04" $
         assertEqual "first extended cf, nothing"
-        (Just (CF.MortgageFlow (L.toDate "20231031") 493.66  0.0 0 0 0 0 0 0 0.00 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20231031") 493.66  0.0 0 0 0 0 0 0 0.00 Nothing Nothing))
         (CF.cfAt poolCf 2)
       ,testCase "case 05" $
         assertEqual "first default from delinq"
-        (Just (CF.MortgageFlow (L.toDate "20240131") 496.69  0.0 0 0 0 1.3 0 1.3 0.000488 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20240131") 499.61  0.0 0 0 0 1.3 0 1.3 0.000952 Nothing Nothing))
         (CF.cfAt poolCf 5)
       ,testCase "case 06" $
         assertEqual "first loss/recovery from default & first back to perf"
-        (Just (CF.MortgageFlow (L.toDate "20240229") 499.66  2.97 0 0 0 1.25 0 1.25 0.000475 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20240229") 496.64  2.97 0 0 0 1.25 0 1.25 0.000480 Nothing Nothing))
         (CF.cfAt poolCf 6)
       -- ,testCase "case 07" $
       --   assertEqual "first loss/recovery from default & first back to perf"
@@ -492,7 +496,7 @@ delinqScheduleCFTest =
       --   (CF.cfAt poolCf 7)
       ,testCase "case with prepay assump" $
         assertEqual "01"
-        (Just (CF.MortgageFlow (L.toDate "20230901") 988.64 0 0 7.02 4.34  0.0 0.0 0.0 0.08 Nothing Nothing))
+        (Just (CF.MortgageDelinqFlow (L.toDate "20230901") 988.64 0 0 7.02 4.34  0.0 0.0 0.0 0.08 Nothing Nothing))
         (CF.cfAt poolCf2 0)
     ]
 
@@ -517,17 +521,29 @@ delinqMortgageTest =
       testCase "" $
         assertEqual "Length of cf"
         5
-        (length txns)
+        (length txns) -- `debug` ("Delinq CF"++show txns)
       ,testCase "first row" $
         assertEqual "delinq = 1"
-        (CF.MortgageFlow (L.toDate "20211101") 159.84 79.12 1.59 0 1.04 0.0 0.0 0.0 0.08 Nothing Nothing)
+        (CF.MortgageDelinqFlow (L.toDate "20211101") 159.84 79.12 1.59 0 1.04 0.0 0.0 0.0 0.08 Nothing Nothing)
         (txns!!0)
+      ,testCase "second row" $
+        assertEqual "with first default/loss/recovery"
+        (CF.MortgageDelinqFlow (L.toDate "20211201") 79.85 79.32 1.06  0 0.67 0.0 0.0 0.0 0.08 Nothing Nothing)
+        (txns!!1)
       ,testCase "last row" $
         assertEqual "with first default/loss/recovery"
-        (CF.MortgageFlow (L.toDate "20220101") 0.0 79.51 0.53  0 0.34 0.31 0.0 0.31 0.08 Nothing Nothing)
+        (CF.MortgageDelinqFlow (L.toDate "20220101") 1.17 79.75 0.53  0 0.34 0.31 0.0 0.31 0.08 Nothing Nothing)
         (txns!!2)
       ,testCase "extend 1st flow" $
         assertEqual "check default"
-        (CF.MortgageFlow (L.toDate "20220201") 0.36 0.36 0.0  0.0 0.0 0.20 0.0 0.2 0.08 Nothing Nothing)
+        (CF.MortgageDelinqFlow (L.toDate "20220201") 0.70 0.47 0.0  0.0 0.0 0.20 0.0 0.2 0.08 Nothing Nothing)
         (txns!!3)
+      -- ,testCase "extend 2st flow" $
+      --   assertEqual "check default"
+      --   (CF.MortgageDelinqFlow (L.toDate "20220201") 1.08 0.36 0.0  0.0 0.0 0.11 0.0 0.11 0.08 Nothing Nothing)
+      --   (txns!!4)
+      -- ,testCase "extend 3st flow" $
+      --   assertEqual "check default"
+      --   (CF.MortgageDelinqFlow (L.toDate "20220201") 1.08 0.36 0.0  0.0 0.0 0.0 0.0 0.0 0.08 Nothing Nothing)
+      --   (txns!!5)
     ]