+++ ChangeLog.md
@@ -1,12 +1,24 @@
 # Changelog for Hastructure
 
+## 0.21.5
+### 2023-10-8
+* ENHANCE: in the revolving buy , now buy amount is no longer a multipler of revolving assets face value
+* FIX: now revolving asset may have remtain term ==  original term
+
+## 0.21.4
+### 2023-9-27
+* ENHANCE: require a new status when defining a deal in `preClosing` stage
+* FIX: fix a bug when reading financial report logs 
+
+
 ## 0.21.3
 ### 2023-9-26
-* NEW: include a default/delinq/loss status map when projecting cashflow
+* NEW: include a `default`/`delinq`/`loss` status map when projecting cashflow
 * NEW: implement `haircut` as extra stress projecting `mortgage` 
 * ENHANCE: include `called` deal status, which will be set when deal was triggered with a clean up call assumption
 * ENHANCE: expose `runAsset` endpoint
 * ENHANCE: expose formula query on `deal status` as well as `trigger status`
+* ENHANCE: add `rampUp` deal status
 * FIX: adjust bond reset date from `cutoff date` to `closing date`
 
 ## 0.21.1